requirements.txt
Flask==3.0.1
elasticsearch==8.17.1
faiss-cpu==1.7.4
numpy==1.23.5